pyproject.toml
[project]
name = "studio"
version = "0.1.0"
description = "A lightweight web studio landing page, and digital tool for small business."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "whitenoise>=6.11.0",
]
